ClenshawCurtisBessel

ClenshawCurtisBessel.BrandersPiessensProblem

ClenshawCurtisBessel.OliverProblem

ClenshawCurtisBessel.Mk_asymptotic

ClenshawCurtisBessel.OliverP

ClenshawCurtisBessel.OliverP

ClenshawCurtisBessel.OliverR

ClenshawCurtisBessel.OliverR

ClenshawCurtisBessel.assembleP

ClenshawCurtisBessel.assembleρ

ClenshawCurtisBessel.bessel_moment_G

ClenshawCurtisBessel.generate_BC

ClenshawCurtisBessel.BrandersPiessensProblem — Type

BrandersPiessensProblem(a, ν)

Construct a boundary value problem from the modified moment recurrence relation.

N and M correspond to the total number of specified initial boundary conditions, and the number specified on the end of the recurrence, respectively. T is the type we're computing on, a is the parameter in the Bessel function Jᵥ(ax), and ν is the order.

ClenshawCurtisBessel.OliverProblem — Type

abstract type OliverProblem{T,N,M} end

Abstract type for constructing a boundary value problem from a recurrence relation. The functions for coefficients and the right side dispatch.

N and M correspond to the total number of specified initial boundary conditions, and the number specified on the end of the recurrence, respectively.

ClenshawCurtisBessel.Mk_asymptotic — Method

Mk_asymptotic(a::T, ν, k) where T

Computes a modified moment using asymptotics as $k \rightarrow \infty$.

ClenshawCurtisBessel.OliverP — Method

OliverP(OP::OliverProblem, s, i)

Computes $P_s(i)$, the coefficient $s$ at solution index $i$.

ClenshawCurtisBessel.OliverP — Method

ClenshawCurtisBessel.OliverP(OP::BrandersPiessensProblem{T}, s, k) where T

Computes the modifed moments from Branders & Piessens 1987.

\[\begin{aligned} \frac{a^2}{16} &M_{k}(a, \nu) + \left[ (k+1)^2 - \nu^2 - \frac{a^2}{4} \right] M_{k + 2}(a, \nu) + \left[ 4 \nu^2 - 2(k+4) + 4 \right] M_{k+3}(a,\nu) \\ &- \left[ 2(k+4)^2 - 6 + 6\nu^2 - \frac{3a^2}{8} \right] M_{k+4}(a,\nu) + (4\nu^2 + 2(k+4) + 4)M_{k+5}(a,\nu) \\ &+ \left[ (k+7)^2 - \nu^2 - \frac{a^2}{4} \right]M_{k+6} + \frac{a^2}{16} M_{k+8}(a,\nu) \quad = 0 \end{aligned}\]

Note that the coefficient indices decrease $P_8, P_7, \cdots, P_0$ as the $M_k$ index increases.

ClenshawCurtisBessel.OliverR — Method

OliverR(OP::OliverProblem, i)

Computes $R(i)$, the right side of the recurrence relation.

ClenshawCurtisBessel.OliverR — Method

OliverR(OP::BrandersPiessensProblem{T}, i)

Computes $R(i)$, the right side of the recurrence relation.

ClenshawCurtisBessel.assembleP — Method

assembleP{T}(OP::OliverProblem{M,N}, a, b) where {M, N}

Generic description of the $\mathbf{P}$ matrix within the linear system of Oliver's method. This in equation 3a of Oliver 1968.

ClenshawCurtisBessel.assembleρ — Method

assembleρ{T}(OP::OliverProblem{M,N}, a, b) where {M, N}

Generic description of the $\vec{\rho}$ vector within the linear system of Oliver's method. This in equation 3b of Oliver 1968.

ClenshawCurtisBessel.bessel_moment_G — Method

bessel_moment_G(λ, ν, a)

Computes the standard Bessel moment $G_{\lambda, \nu}(a)$ defined as

\[G_{\lambda, \nu}(a) = \int_0^a t^{\lambda} J_{\nu}(t)\,dt.\]

ClenshawCurtisBessel.generate_BC — Method

generate_BC(OP::BrandersPiessensProblem{T,8,2}, kmax::Int) where T

Generates the boundary conditions required for Oliver's method to solve the boundary value problem posed Branders & Piessens.

Arguments:

OP::BrandersPiessensProblem{T,8,2}: problem parameters

kmax: maximum order $k$ to compute.

Returns:

Tuple{Int,OffsetVector}: the maximum index, and a sparse offset array with the boundaries.